fix(adapter-next): don't purge user cache on simulator update
Previously when using the Next.js adapter and the Server Component
`<SliceSimulator />`, 'simulator-slices' updates from the
simulator would call `revalidatePath` from Next.js in order to
re-render the `<SliceSimulator />` component with the updated
search params. A 'simulator-slices' update occurs every time a
user makes a change in the simulator, for example a keystroke or
slice field change.

This is a problem when slices contain data fetches. Next.js has a
Data Cache, which allows a user to specify a cache time on `fetch`
requests, or use api's like `unstable_cache` to cache the result
of an operation. `revalidatePath` purges the Next.js Data Cache
for that path, which means that if a slice is fetching data from
an external source, that fetch will skip the Data Cache and hit
origin.

This is problematic if the fetch or operation is expensive. Users
could have wrapped expensive operations inside `unstable_cache`,
or specified a cache time using `revalidate`, and those values
would not be respected because of the `revalidatePath` call.
Although this call is throttled to a maximum of 1 every 300
milliseconds, this could very quickly add up to a significant
number. These calls are not made in user code either, they are
made within the slice simulator, so this could result in a lot of
unexpected and unwanted calls to origin for users.

This commit fixes this by replacing `revalidatePath` with
`router.refresh`. Since the simulator state is in search params,
and therefore not in Data Cache, there is no need to purge the
cache. We just need to re-run the Server Component, which
`router.refresh` does.